requirements.txt
Django==1.7.1
Pillow==2.6.1
argparse==1.2.1
django-bootstrap-form==3.1
pyflakes==0.8.1
sorl-thumbnail==12.1c
wsgiref==0.1.2
